added  interface to types/index.ts
